create new components for header and homepage; restyle announcement
